+++ backend/src/main/java/groom/backend/interfaces/coupon/CouponController.java
@@ -33,7 +33,7 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
 
-@Tag(name = "Coupon", description = "쿠폰 발급 및 사용 관련 API")
+@Tag(name = "Coupon", description = "쿠폰 발급 및 사용, 조회 API")
 @Validated
 @Slf4j
 @RestController
@@ -42,17 +42,6 @@ import java.util.List;
 public class CouponController {
   private final CouponCommonService couponCommonService;
   private final CouponIssueService couponIssueService;
-
-  @PostMapping
-  @Operation(summary = "쿠폰 생성", description = "지정된 값으로 쿠폰을 생성합니다. 할인 정책을 적용시 해당 정책의 수치는 변경할 수 없습니다.")
-  @ApiResponses(value = {
-          @ApiResponse(responseCode = "201", description = "Created",
-                  content = {@Content(schema = @Schema(implementation = CouponResponse.class))}),
-  })
-  public ResponseEntity<CouponResponse> createCoupon(@Validated @RequestBody CouponCreateRequest request) {
-    CouponResponse response = couponCommonService.createCoupon(request);
-    return ResponseEntity.status(HttpStatus.CREATED).body(response);
-  }
 
   @GetMapping("/{coupon_id}")
   @Operation(summary = "단일 쿠폰 조회", description = "지정된 id의 쿠폰을 조회합니다.")
@@ -91,50 +80,6 @@ public class CouponController {
           @PageableDefault(size = 10) Pageable pageable) {
     Page<CouponResponse> response = couponCommonService.searchCoupon(condition, pageable);
     return ResponseEntity.ok(response);
-  }
-
-  @Operation(
-          summary = "쿠폰 수정",
-          description = "쿠폰 ID를 기반으로 쿠폰 정보를 수정합니다."
-  )
-  @ApiResponses({
-          @ApiResponse(responseCode = "200", description = "수정 성공",
-                  content = @Content(schema = @Schema(implementation = CouponResponse.class))),
-          @ApiResponse(responseCode = "404", description = "수정 실패",
-                  content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-  })
-  @PutMapping("/{coupon_id}")
-  public ResponseEntity<CouponResponse> updateCoupon(
-          @Parameter(description = "쿠폰 ID", example = "1")
-          @PathVariable("coupon_id") Long couponId,
-          @RequestBody CouponUpdateRequest request) {
-    CouponResponse response = couponCommonService.updateCoupon(couponId, request);
-    if (response == null) {
-      return ResponseEntity.notFound().build();
-    }
-    return ResponseEntity.ok(response);
-  }
-
-  @Operation(
-          summary = "쿠폰 삭제",
-          description = "쿠폰 ID를 기반으로 쿠폰을 삭제합니다."
-  )
-  @ApiResponses({
-          @ApiResponse(responseCode = "204", description = "삭제 성공"),
-          @ApiResponse(responseCode = "404", description = "삭제 실패"),
-//          TODO : 현재 Coupon 삭제 시 실패 요청은 메시지를 던지지 않으며, 컨트롤러에서 직접 검증함.
-//          @ApiResponse(responseCode = "404", description = "요청한 리소스를 찾을 수 없습니다.",
-//                  content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-  })
-  @DeleteMapping("/{coupon_id}")
-  public ResponseEntity<Void> deleteCoupon(@PathVariable("coupon_id") Long couponId) {
-    Boolean result = couponCommonService.deleteCoupon(couponId);
-    if (!result) {
-      // 삭제 실패 (존재하지 않음)
-      return ResponseEntity.notFound().build();
-    }
-    // 삭제 성공
-    return ResponseEntity.noContent().build();
   }
 
   @Operation(
